fix(models): move models to api-schema
