+++ apps/web/src/api-schema/common/model.ts
@@ -1,0 +1,9 @@
+/* eslint-disable camelcase */
+export interface Tags {
+    [key: string]: any;
+}
+
+export interface TimeStamp {
+    seconds: string;
+    nanos?: number;
+}
