+++ apps/web/src/services/administration/type.ts
@@ -1,4 +1,4 @@
-import type { TimeStamp } from '@/models';
+import type { TimeStamp } from '@/api-schema/common/model';
 
 import type { RoleData } from '@/services/administration/iam/role/type';
 import type { ProjectGroupInfo, ProjectModel } from '@/services/project/type';
