+++ apps/web/src/services/asset-inventory/cloud-service/cloud-service-detail/modules/CloudServiceDetail.vue
@@ -17,8 +17,8 @@ import { QueryHelper } from '@cloudforet/core-lib/query';
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import { ApiQueryHelper } from '@cloudforet/core-lib/space-connector/helper';
 
-import type { ExportParameter } from '@/models/export';
-import { QueryType } from '@/models/export';
+import type { ExportParameter } from '@/api-schema/common/api-verbs/export';
+import { QueryType } from '@/api-schema/common/api-verbs/export';
 import { store } from '@/store';
 import { i18n } from '@/translations';
