+++ apps/web/src/services/asset-inventory/cloud-service/cloud-service-detail/modules/ExcelExportOptionModal.vue
@@ -13,8 +13,8 @@ import type { ConsoleFilter } from '@cloudforet/core-lib/query/type';
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import { ApiQueryHelper } from '@cloudforet/core-lib/space-connector/helper';
 
-import type { ExportOption, ExportParameter } from '@/models/export';
-import { QueryType } from '@/models/export';
+import type { ExportOption, ExportParameter } from '@/api-schema/common/api-verbs/export';
+import { QueryType } from '@/api-schema/common/api-verbs/export';
 import { store } from '@/store';
 import { i18n } from '@/translations';
