+++ apps/web/src/services/asset-inventory/cloud-service/modules/CloudServiceToolbox.vue
@@ -13,8 +13,8 @@ import { QueryHelper } from '@cloudforet/core-lib/query';
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import { ApiQueryHelper } from '@cloudforet/core-lib/space-connector/helper';
 
-import type { ExportParameter, ExportOption } from '@/models/export/index';
-import { QueryType } from '@/models/export/index';
+import type { ExportParameter, ExportOption } from '@/api-schema/common/api-verbs/export';
+import { QueryType } from '@/api-schema/common/api-verbs/export';
 import { store } from '@/store';
 
 import type { ProviderReferenceMap } from '@/store/modules/reference/provider/type';
